+++ integration/testdata/exit-code-propagation/Dockerfile_exit_code_propagation
